Updated ProductController

+++ src/main/java/com/tts/ecommerce/controller/ProductController.java
@@ -1,5 +1,7 @@
 package com.tts.ecommerce.controller;
 
+import com.tts.ecommerce.model.Product;
+import com.tts.ecommerce.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -17,7 +19,7 @@ public class ProductController {
   ProductService productService;
 
   @GetMapping(value = "/product/{id}")
-  public String show(@PathVariable int id, Model model) {
+  public String show(@PathVariable long id, Model model) {
     Product product = productService.findById(id);
     model.addAttribute(product);
     return "product";
